+++ crates/cubecl-farm/Cargo.toml
@@ -18,9 +18,7 @@ cubecl-runtime = { path = "../cubecl-runtime", version = "0.6.0", default-featur
 cubecl-cuda = { path = "../cubecl-cuda", version = "0.6.0" }
 bytemuck = { workspace = true }
 cudarc = { version = "0.16.3", features = [
-  "std",
   "driver",
-  "cuda-version-from-build-system",
-  "dynamic-loading",
   "nccl",
 ], default-features = false }
+crossbeam-channel = "0.5.15"
